tests/wpt/tests/dom/nodes/Document-contentType/contentType/xhr_responseType_document.html

<!DOCTYPE html>
<title>XHR - retrieve HTML document: document.contentType === 'application/xml'</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var xhr = new XMLHttpRequest();
  xhr.open("GET", "../resources/blob.xml");
  xhr.responseType = "document";
  xhr.onload = this.step_func_done(function(response) {
    assert_equals(xhr.readyState, 4);
    assert_equals(xhr.status, 200);
    assert_equals(xhr.responseXML.contentType, "application/xml");
  });
  xhr.send(null);
});
</script>
